[OpenACC] add isEffectivelySerial() to acc.parallel and acc.kernels
Mirror the isEffectivelySerial() function in acc.compute_region to
acc.parallel and acc.kernel so this check is unified anywhere it needs
to be used. Should be NFC
